requirements.txt
streamlit
pandas
google-generativeai
tabulate